refactor(backend): padronizar enums, ajustar regras de negocio e fluxo financeiro

+++ hungo-spring/src/main/java/br/com/halotec/hungospring/dto/ItemPedidoDTO.java
@@ -1,23 +1,25 @@
 package br.com.halotec.hungospring.dto;
 
+import java.math.BigDecimal;
+
 public class ItemPedidoDTO {
     private Long id;
     private Long produtoId;
     private Integer quantidade;
-    private Float total;
+    private BigDecimal total;
     private String statusItem;
     private String motivoCancelamento;
 
     public ItemPedidoDTO() {}
 
-    public ItemPedidoDTO(Long id, Long produtoId, Integer quantidade, Float total) {
+    public ItemPedidoDTO(Long id, Long produtoId, Integer quantidade, BigDecimal total) {
         this.id = id;
         this.produtoId = produtoId;
         this.quantidade = quantidade;
         this.total = total;
     }
 
-    public ItemPedidoDTO(Long id, Long produtoId, Integer quantidade, Float total, String statusItem, String motivoCancelamento) {
+    public ItemPedidoDTO(Long id, Long produtoId, Integer quantidade, BigDecimal total, String statusItem, String motivoCancelamento) {
         this.id = id;
         this.produtoId = produtoId;
         this.quantidade = quantidade;
@@ -32,8 +34,8 @@ public class ItemPedidoDTO {
     public void setProdutoId(Long produtoId) { this.produtoId = produtoId; }
     public Integer getQuantidade() { return quantidade != null ? quantidade : 0; }
     public void setQuantidade(Integer quantidade) { this.quantidade = quantidade; }
-    public Float getTotal() { return total; }
-    public void setTotal(Float total) { this.total = total; }
+    public BigDecimal getTotal() { return total; }
+    public void setTotal(BigDecimal total) { this.total = total; }
     public String getStatusItem() { return statusItem; }
     public void setStatusItem(String statusItem) { this.statusItem = statusItem; }
     public String getMotivoCancelamento() { return motivoCancelamento; }
