+++ hungo-spring/src/main/java/br/com/halotec/hungospring/dto/ProdutoDTO.java
@@ -1,9 +1,11 @@
 package br.com.halotec.hungospring.dto;
 
+import java.math.BigDecimal;
+
 public class ProdutoDTO {
     private Long id;
     private String nome;
-    private Float preco;
+    private BigDecimal preco;
     private Long categoriaId;
     private Boolean tipo;
     private Boolean favorito = false;
@@ -14,8 +16,8 @@ public class ProdutoDTO {
     public void setId(Long id) { this.id = id; }
     public String getNome() { return nome; }
     public void setNome(String nome) { this.nome = nome; }
-    public Float getPreco() { return preco; }
-    public void setPreco(Float preco) { this.preco = preco; }
+    public BigDecimal getPreco() { return preco; }
+    public void setPreco(BigDecimal preco) { this.preco = preco; }
     public Long getCategoriaId() { return categoriaId; }
     public void setCategoriaId(Long categoriaId) { this.categoriaId = categoriaId; }
     public Boolean getTipo() { return tipo; }
