+++ hungo-spring/src/main/java/br/com/halotec/hungospring/repository/ItemPedidoRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Long> {
@@ -17,10 +18,10 @@ public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Long> {
     void deleteByPedidoId(@Param("pedidoId") Long pedidoId);
 
     @Query("SELECT COALESCE(SUM(i.total), 0.0) FROM ItemPedido i WHERE i.pedido.venda.id = :vendaId AND (i.statusItem IS NULL OR UPPER(i.statusItem) <> 'CANCELADO') AND (i.pedido.statusPedido IS NULL OR UPPER(i.pedido.statusPedido) <> 'CANCELADO')")
-    Float somarTotalPorVendaId(@Param("vendaId") Long vendaId);
+    BigDecimal somarTotalPorVendaId(@Param("vendaId") Long vendaId);
 
     @Query("SELECT COALESCE(SUM(i.total), 0.0) FROM ItemPedido i WHERE i.pedido.venda.id = :vendaId AND (UPPER(i.pedido.statusPedido) = 'CONCLUÍDO' OR UPPER(i.pedido.statusPedido) = 'CONCLUIDO') AND (i.statusItem IS NULL OR UPPER(i.statusItem) <> 'CANCELADO')")
-    Float somarTotalConcluidoPorVendaId(@Param("vendaId") Long vendaId);
+    BigDecimal somarTotalConcluidoPorVendaId(@Param("vendaId") Long vendaId);
 
     List<ItemPedido> findByPedidoId(Long pedidoId);
     List<ItemPedido> findByPedidoVendaId(Long vendaId);
