+++ hungo-spring/src/main/java/br/com/halotec/hungospring/repository/VendaRepository.java
@@ -5,13 +5,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
 public interface VendaRepository extends JpaRepository<Venda, Long> {
     List<Venda> findByDataFimVendaIsNull();
     List<Venda> findByDataFimVendaIsNotNull();
-    List<Venda> findByValorPagoGreaterThanAndDataFimVendaIsNull(Float valorPago);
+    List<Venda> findByValorPagoGreaterThanAndDataFimVendaIsNull(BigDecimal valorPago);
     List<Venda> findByMesaIdAndDataFimVendaIsNull(Long mesaId);
     List<Venda> findByClienteIdAndStatusPagamentoIn(Long clienteId, List<String> statusPagamentos);
     List<Venda> findByStatusPagamentoIn(List<String> statusPagamentos);
